requirements.txt
# =====================================================
# MINGUS APPLICATION - PRODUCTION DEPENDENCIES
# =====================================================
# For development dependencies, see requirements-dev.txt

# =====================================================
# CORE FRAMEWORK
# =====================================================
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3

# =====================================================
# DATABASE AND ORM
# =====================================================
alembic==1.13.3
greenlet==3.2.2
psycopg2-binary==2.9.10
SQLAlchemy==2.0.41
sqlalchemy-searchable==2.1.0
sqlalchemy-utils==0.41.2

# =====================================================
# AUTHENTICATION AND SECURITY
# =====================================================
bcrypt==4.1.2
cryptography==42.0.8
django-encrypted-model-fields==0.6.3
passlib==1.7.4
PyJWT==2.8.0
Flask-Talisman==1.1.0
Flask-Limiter==3.5.0
Flask-WTF==1.2.1
WTForms==3.1.2

# =====================================================
# SUPABASE INTEGRATION
# =====================================================
gotrue==2.11.4
postgrest==1.0.2
realtime==2.4.3
storage3==0.11.3
supabase==2.15.2
supafunc==0.9.4

# =====================================================
# PAYMENT PROCESSING
# =====================================================
stripe==12.3.0

# =====================================================
# FINANCIAL DATA INTEGRATION
# =====================================================
plaid-python==18.0.0

# =====================================================
# COMMUNICATION SERVICES
# =====================================================
aiohttp==3.12.6
httpx==0.28.1
resend==0.8.0
twilio==9.3.0
websockets==14.2

# =====================================================
# BACKGROUND TASKS AND CACHING
# =====================================================
celery==5.3.4
redis==6.2.0
schedule==1.2.0
flower>=2.0.1

# =====================================================
# HTTP AND NETWORKING
# =====================================================
aiohappyeyeballs==2.6.1
aiosignal==1.3.2
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
frozenlist==1.6.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
hyperframe==6.1.0
idna==3.10
multidict==6.4.4
requests==2.31.0
sniffio==1.3.1
urllib3==2.4.0
yarl==1.20.0

# =====================================================
# DATA PROCESSING AND ANALYSIS
# =====================================================
numpy==1.26.4
pandas==2.2.1
python-dateutil==2.8.2
pytz==2025.2

# =====================================================
# ARTICLE LIBRARY DEPENDENCIES
# =====================================================
# AI and Natural Language Processing
openai>=1.0.0
nltk>=3.8.1
textstat>=0.7.3

# Web Scraping and Content Extraction
newspaper3k>=0.2.8
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.10

# Search and Indexing
whoosh>=2.7.4
elasticsearch>=8.0.0
elasticsearch-dsl>=8.0.0

# Content Processing
python-magic>=0.4.27
chardet>=5.0.0
langdetect>=1.0.9

# =====================================================
# CONFIGURATION AND ENVIRONMENT
# =====================================================
python-dotenv==1.0.1

# =====================================================
# MONITORING AND HEALTH CHECKS
# =====================================================
psutil==5.9.6
prometheus-client==0.19.0
PyYAML==6.0.2

# =====================================================
# MONITORING AND LOGGING
# =====================================================
sentry-sdk[flask]>=1.32.0
structlog>=23.1.0

# =====================================================
# VALIDATION AND SERIALIZATION
# =====================================================
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
pydantic==2.11.5
pydantic_core==2.33.2
referencing==0.36.2
rpds-py==0.25.1

# =====================================================
# PRODUCTION SERVER
# =====================================================
gunicorn==21.2.0

# =====================================================
# UTILITIES
# =====================================================
blinker==1.9.0
click==8.2.1
deprecation==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mistune==3.1.3
StrEnum==0.4.15
tzdata==2025.2
# =====================================================
# DEVELOPMENT AND TESTING
# =====================================================
pytest>=7.4.0
pytest-flask>=1.2.0
pytest-cov>=4.1.0
pytest-html>=3.2.0
factory-boy>=3.3.0
selenium>=4.0.0
webdriver-manager>=3.8.0
